requirements.txt
annotated-types==0.6.0
anyio==4.2.0
cachetools==5.3.2
certifi==2024.2.2
cffi==1.16.0
cfgv==3.4.0
charset-normalizer==3.3.2
click==8.1.7
cryptography==42.0.2
distlib==0.3.8
exceptiongroup==1.2.0
fastapi==0.109.2
fastapi-sessions==0.3.2
filelock==3.13.1
google-api-core==2.17.0
google-api-python-client==2.117.0
google-auth==2.27.0
google-auth-httplib2==0.2.0
google-auth-oauthlib==1.2.0
googleapis-common-protos==1.62.0
gunicorn==21.2.0
h11==0.14.0
httpcore==1.0.2
httplib2==0.22.0
httptools==0.6.1
httpx==0.26.0
httpx-oauth==0.13.2
identify==2.5.34
idna==3.6
install==1.3.5
itsdangerous==2.1.2
Jinja2==3.1.3
MarkupSafe==2.1.5
nodeenv==1.8.0
oauthlib==3.2.2
packaging==23.2
pillow==10.2.0
pip-autoremove==0.10.0
platformdirs==4.2.0
pre-commit==3.6.1
protobuf==4.25.2
pyasn1==0.5.1
pyasn1-modules==0.3.0
pycparser==2.21
pydantic==2.6.1
pydantic_core==2.16.2
pyparsing==3.1.1
python-dotenv==1.0.1
python-multipart==0.0.9
PyYAML==6.0.1
requests==2.31.0
requests-oauthlib==1.3.1
rsa==4.9
sniffio==1.3.0
starlette==0.36.3
typing_extensions==4.9.0
uritemplate==4.1.1
urllib3==2.2.0
uvicorn==0.27.1
virtualenv==20.25.0
watchfiles==0.21.0
websockets==12.0
